+++ Sources/MapboxMaps/Foundation/MapInitOptions.swift
@@ -1,6 +1,10 @@
 import Foundation
 
+/// A protocol used to provide ``MapInitOptions`` when initializing a ``MapView`` with a Storyboard or
+/// a nib.
 @objc public protocol MapInitOptionsProvider {
+    /// A method to be called when ``MapView`` needs initialization options
+    /// - Returns: Initializations options for the ``MapView``.
     func mapInitOptions() -> MapInitOptions
 }
 
@@ -16,10 +20,10 @@ public final class MapInitOptions: NSObject {
     /// Associated `MapOptions`
     public let mapOptions: MapOptions
 
-    /// Default style URI for initializing the map
+    /// Style URI for initializing the map. Defaults to Mapbox Streets.
     public let styleURI: StyleURI?
 
-    /// Default camera options for initializing the map
+    /// Camera options for initializing the map. CameraOptions default to 0.0 for each value.
     public let cameraOptions: CameraOptions?
 
     /// Initializer. The default initializer, i.e. `MapInitOptions()` will use