+++ Sources/MapboxMaps/Foundation/MapView.swift
@@ -411,7 +411,9 @@ open class MapView: UIView {
         location = dependencyProvider.makeLocationManager(
             locationProducer: locationProducer,
             interpolatedLocationProducer: interpolatedLocationProducer,
-            style: mapboxMap.style)
+            style: mapboxMap.style,
+            mapboxMap: mapboxMap,
+            displayLinkCoordinator: self)
 
         // Initialize/Configure annotations orchestrator
         annotations = AnnotationOrchestrator(
@@ -557,7 +559,11 @@ open class MapView: UIView {
         mapboxMap.size = bounds.size
     }
 
+    @_spi(Metrics) public var metricsReporter: MapViewMetricsReporter?
+
     private func updateFromDisplayLink(displayLink: CADisplayLink) {
+        metricsReporter?.beforeDisplayLinkCallback(displayLink: displayLink)
+        defer { metricsReporter?.afterDisplayLinkCallback(displayLink: displayLink) }
         if window == nil {
             return
         }
@@ -572,7 +578,9 @@ open class MapView: UIView {
 
         if needsDisplayRefresh {
             needsDisplayRefresh = false
+            metricsReporter?.beforeMetalViewDrawCallback(metalView: metalView)
             metalView?.draw()
+            metricsReporter?.afterMetalViewDrawCallback(metalView: metalView)
         }
     }