+++ Sources/MapboxMaps/Ornaments/OrnamentsManager.swift
@@ -1,16 +1,30 @@
 import UIKit
 
+/// Options used to configure the corner position of an ornament
 public enum OrnamentPosition: String, Equatable {
     // Clockwise from top left
+    @available(*, deprecated, renamed: "topLeading")
     case topLeft
+    @available(*, deprecated, renamed: "topTrailing")
     case topRight
+    @available(*, deprecated, renamed: "bottomTrailing")
     case bottomRight
+    @available(*, deprecated, renamed: "bottomLeading")
     case bottomLeft
+
+    case topLeading
+    case topTrailing
+    case bottomLeading
+    case bottomTrailing
 }
 
+/// Options used to configure the visibility of an ornament
 public enum OrnamentVisibility: String, Equatable {
+    /// Shows the ornament when relevant
     case adaptive
+    /// Hides the ornament
     case hidden
+    /// Shows the ornament
     case visible
 }
 
@@ -19,9 +33,10 @@ internal struct Ornaments {
     static let telemetryURL = "https://www.mapbox.com/telemetry/"
 }
 
+/// APIs for managing map ornaments
 public class OrnamentsManager: NSObject {
 
-    /// The `OrnamentOptions` object that is used to set up and update the required ornaments on the map.
+    /// The ``OrnamentOptions`` object that is used to set up and update the required ornaments on the map.
     public var options: OrnamentOptions {
         didSet {
             updateOrnaments()
@@ -44,7 +59,7 @@ public class OrnamentsManager: NSObject {
 
     /// The view for the compass ornament. This view can be used to position other views relative to the
     /// compass ornament, but it should not be manipulated. Use ``OrnamentOptions/compass`` to
-    /// configure the compass presentation if customization is needed..
+    /// configure the compass presentation if customization is needed.
     public var compassView: UIView {
         return _compassView
     }
@@ -52,7 +67,7 @@ public class OrnamentsManager: NSObject {
     /// The view for the attribution button ornament. This view can be used to position other views relative
     /// to the attribution button ornament, but it should not be manipulated. Use
     /// ``OrnamentOptions/attributionButton`` to configure the attribution button presentation
-    /// if customization is needed..
+    /// if customization is needed.
     public var attributionButton: UIView {
         return _attributionButton
     }
@@ -162,6 +177,9 @@ public class OrnamentsManager: NSObject {
                                                        margins: options.attributionButton.margins)
         constraints.append(contentsOf: attributionButtonConstraints)
 
+        // Update the image of compass
+        _compassView.updateImage(image: options.compass.image)
+
         // Activate new constraints
         NSLayoutConstraint.activate(constraints)
 
@@ -191,6 +209,22 @@ public class OrnamentsManager: NSObject {
         case .bottomRight:
             return [
                 view.rightAnchor.constraint(equalTo: layoutGuide.rightAnchor, constant: -margins.x),
+                view.bottomAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -margins.y)]
+        case .topLeading:
+            return [
+                view.leadingAnchor.constraint(equalTo: layoutGuide.leadingAnchor, constant: margins.x),
+                view.topAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: margins.y)]
+        case .topTrailing:
+            return  [
+                view.trailingAnchor.constraint(equalTo: layoutGuide.trailingAnchor, constant: -margins.x),
+                view.topAnchor.constraint(equalTo: layoutGuide.topAnchor, constant: margins.y)]
+        case .bottomLeading:
+            return [
+                view.leadingAnchor.constraint(equalTo: layoutGuide.leadingAnchor, constant: margins.x),
+                view.bottomAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -margins.y)]
+        case .bottomTrailing:
+            return [
+                view.trailingAnchor.constraint(equalTo: layoutGuide.trailingAnchor, constant: -margins.x),
                 view.bottomAnchor.constraint(equalTo: layoutGuide.bottomAnchor, constant: -margins.y)]
         }
     }